[ new ] Language with nested scope
At the moment SyntaxWithBinding does not handle well declarations
with nested use of the scope transformer such as:

> data  YF r s a =
>      A (r s a) (r s a)
>   |  Y (s (s (r s)) a)

where the intended meaning of `Y` is:

> y :: ((a -> b) -> (a -> b)) -> (a -> b)
> y f x = f (y f) x

Indeed for these languages it is impossible to implement reindex.
Here we have an example of an "exploded" version of `YF` where
`Y` is morally equivalent to `Y2 . Y1`. The expected language can
be recovered by using a pattern synonym and hiding these two
constructors.

We can even implement normalisation by evaluation for such a language
but it is slightly annoying: when dealing with a term whose denotation
should be `y f x`, rather than applying a `f` to `Y f` itself and its
argument, we have to take the fixpoint of the value packaged by `Y2`,
inspect the body of the resulting model value to peel off `Y1` and then
apply the argument. There is no way for Haskell to know that we have
covered all possible cases but as long as the user only has direct
access to the pattern synonym `Y`, it is indeed the case.

It would obviously be nicer to be able to have this invariant enforced
statically.