[ cleanup ] Removing the complexity added by CsF
It's nice to know we are able to do this sort of things however
it clutters massively the presentation, and makes it hard if not
impossible to write good specifications (e.g. reindex fusion).

Allowing more liberal notions of scoping rules is now fully
relegated to the 'future work' section.